Big commit.
Added statistics library, looking for some places to use it, maybe in T student and log-normal.
Double slider fixed and working properly.
Sidebar on mobile fixed and working.
Little issue with state and distro changes solved.
Created and styled the Result area with a ripple button component.
Extra dara refactored and moved to the top